move attributesWithValuesForCaseClass to ProductUtil, add scaladoc
